+++ CAttaque.h
@@ -8,28 +8,28 @@ class CAttaque
 {
 protected:
 	std::string m_nom;
-	short m_type;
-	short m_nu; //nombre d'utilisation de l'attaque
-	short m_pAttaque; //puissance d'attaque
+	int m_type;
+	int m_nu; //nombre d'utilisation de l'attaque
+	int m_pAttaque; //puissance d'attaque
 	float m_pEchec; //probabilité d'échec
 	bool m_cs;//s'agit t'il d'une capacité spéciale ?
 
 public:
 	CAttaque();
-	CAttaque(std::string nom, short type, short nu, short pa, float pe);
+	CAttaque(std::string nom, int type, int nu, int pa, float pe);
 	~CAttaque();
 
 	std::string getNom();
-	short getType();
-	short getNU();
-	short getPA();
+	int getType();
+	int getNU();
+	int getPA();
 	float getPE();
 	bool getCS();
 
 	void setNom(std::string nom);
-	void setType(short type);
-	void setNU(short nu);
-	void setPA(short pa);
+	void setType(int type);
+	void setNU(int nu);
+	void setPA(int pa);
 	void setPE(float pe);
 	void setCS(bool cs);
 };
